[MiqReport::Generator] Avoid polymorphic to Rbac
This changes what is passed to Rbac from MiqReport::Generator to avoid
passing keys for includes/references that are polymorphic in nature.

This means that there might be a bit of a performance degradation on
some reports, but for now stability is being favored over performance.

Follow up:  For the keys that are skipped, a MiqPreload should probably
be called to avoid N+1 calls, but unsure where the best for this would
be.