[GitWorktree#checkout_at] Pass current commit_sha
This fixes a bug where if a user specified a particular branch, it would
not be respected by the system, and would always checkout origin/master
for the parent repo.

This was not a problem for submodules, since the `@commit_sha` was
already being passed in for the clone of those repos.