Only queue ansible runner ops when not in pods
This commit adds an intermediate method (#route_signal) to determine
if a call should be queued or not.

When running in containers, each generic worker is a separate container
so we can't queue anything between checking out the playbook
repository and cleaning it up. If we do, it might end up executing
on a container that doesn't have the repo checked out or isn't
running the ansible runner process.

We want to continue queueing these operations on appliances as the
previous reasoning doesn't apply (we will always queue for a worker
on the same server) and we still need to handle ansible playbooks
that might run longer than the timeout for a single queue message.

For now these long-running playbooks won't have a solution on pods,
but shorter ones will work.