Fix Full Pathname Stored in DB
The filesystem attribute in the XML produced by SSA currently contains the
search path specified by the user in the scan profile.  Since we can now add
glob characters to directory names with the addition of
https://github.com/ManageIQ/manageiq-smartstate/pull/114,
an issue arises in that the Filesystem class which parses the XML will use the
search path plus the basename of the discovered file to build the full pathname,
and store it in the DB.  This pathname can now sometimes contain a glob character,
which does not provide a full path.

Conveniently, the XML already contained an fqname element which already contained the
*real* fully qualified, expanded pathname.  If the fqname element exists in the XML,
we'll use that for the DB element.