Fix a bug when running Rbac on VimPerformanceDaily
`Rbac.filtered(VimPerformanceDaily)` rendered the following:

```
     Failure/Error: return klass.name[0..-12].constantize.base_class
     
     NameError:
       uninitialized constant VimPer
```

The problem was that it was saying that `VimPerformanceTag` has an associated with a class.

`VmPerformance` and `HostPerformance` have associated classes (`Vm` and `Host` respectively).
This code derives the associated code from the class name by
removing `"performance"` (11 letters) from the end.

But `VimPerformanceTag` is not associated with a model, so
stripping off the last 11 letters (which aren't the word performance)