[Ansible::Runner] Fix --ask-become-method for machine credentials
The `ansible-plabook` command has an option to ask for the password that
will when escalating privileges on the target host, which is
`--ask-become-method`.

`ansible-runner` uses the `env/passwords` file as a collection of STDOUT
regexps to match against (keys), in which then should send the
corresponding password (values) to STDIN as input.

The problem exists when a machine credential exists that doesn't have a
password, so `env/passwords` isn't populated with a match for the prompt
"BECOME Password: ", and hangs until a timeout is processed.

This fixes moves the logic for determining if `--ask-become-password`
should be included in the `ansible-playbook` commandline ops to the
`lib/ansible/runner/credential/machine_credential.rb` and makes it
conditional on the presence of a `become_password` in the auth record.