Ensure ems workers are killed by their server/orchestrator pod
Fixes #20288

Previously, direct calls to ems#destroy would assume you're calling it local
to each of the ems's workers and would fail to find the pid if not local.  Additionally,
in pods, only the orchestrator pod of the worker has permissions to kill the pod
so this would fail with permission errors such as:

deployments.apps "1-xyz-event-catcher-1" is forbidden: User "abc" cannot patch resource "deployments"
in API group "apps" in the namespace "123" for PATCH https:...]

The ems.destroy_queue method calls _queue_task from the AsyncDeleteMixin, which doesn't specify the server_guid or queue_name
so a UI request to delete the ems COULD be initiated in a UI appliance and picked up by the same appliance,
which isn't where the ems's worker processes are running, and would ultimately call kill on each workers that don't exist locally.

Now, we queue the worker's kill method for the queue_name 'miq_server' so it's handled by the server "process" in appliances
or orchestrator in pods and server_guid of the worker's server as an ems's workers can be on different servers.