[EmbeddedAnsible] Fix edit credentials
There was a bug when editing credentials for EmbeddedAnsible where if
you edited an attribute and there are other attributes on the particular
credential class' params_to_attributes method which have to be
converted, those previously would be `nil`'d out.

This fixes that, and modifies tests to confirm that this is now working.