Consolidate monitor_pods to loop over: reset, get initial, get updates
Bonus:  Now, all modifications to the shared resource current_pods happen in the
single monitor thread.  The callers outside of this thread either initialize
current_pods if not already done or read.

Before, it was split apart.  If an error occurred in updates
(watch_for_pod_events), we'd break out of that method early and not repeat the
process cleanly.

Now, we expect errors may come so we do:
1) reset
2) collect initial
3) collect updates
in an endless loop.