[Ansible::Runner] wait on artifacts/ to exist
Before returning a result, wait on the `artifacts/` directory to be
created before returning a response object.

This will ensure that calling `.running?` will return `false` properly
when the `ansible-runner` has ended, and not before it has even had a
chance to start.

Thanks for Jason for the find on this bug.