drop migration from models
The code looks for cloud volumes and updates their value based upon the ems.
This seems like a data migration that would be better suited in db:migrations

But what has me confused is how we have cloud volumes associated with our ems when we
are in the process of creating our ems and providers.

The cider_manager references this object and needs this object (the parent namager)
to be saved before it can be saved.
This is confusing because we are in the before create method and we don't have an
id yet.

So before you create this object we want to save the child object that needs us
to be saved first so it can get a reference to our id. The infinite loop is averted
thanks to rails being smart, but something is wrong with the logic.

As far as I can tell, the zone is not valid while saving records at this time, so
the save and reload do nothing.