[rails6][git_repository_spec.rb] Use .local_miq_server
In Rails 6, it seems like a validation callback for nested relations is
causing a `GitRepository.save!` to fail.  In this case, a MiqQueue call
validation that checks that the `Zone` that is being provided via the
`options` to queue the work on actually exists.

This seems to have not been respected properly in the past, so this fix
just simply switches to using actually `MiqServer`/`Zone` records
instead of trying to stub them out.

Works in both Rails 5.2 and Rails 6.0