Deprecate :embedded_automation_manager_ansible factory
This factory is the incorrect way for generating a manager for
EmbeddedAnsible, since a lot of what is necessary for the Manager
requires the Provider, and the provider has it's own mechanism for
ensuring a AutomationManager exists for it.

See commit 31a2f329b8809b76c1ef8fce572edde9917973de

As a result, it is better to prevent the usage of this factory, which in
this case, we do a lazy loaded `raise` that will get called when the
factory attempts to assign the `provider` value.

Instead, users should create a Provider record using the
`:provider_embedded_ansible` factory, and call to the
`automation_manager` associated with that record.