Write systemd service files on startup not in sync_workers
When the systemd worker option was first introduced as a prototype there
was a setting that would control if systemd or the normal forking worker
mode was used.  This setting was per worker, and was done in
sync_workers so that you could change the setting "on-the-fly" without
restarting the entire evmserverd process.  This made it much easier to
try out using systemd on one or two workers very quickly.

Now that systemd is the default for appliances the ability to change
this dynamically isn't very helpful, and writing out systemd unit files
during sync_workers is just a lot of overhead.

Moving the writing of systemd unit files to miq_server startup removes
the overhead per-sync_workers-loop and allows us to update the contents
of the files without checking .exist? (which was done due to the
frequency with which it was being run).