[MiqPglogical] Add ConnectionHandling
In addition to some more refactoring, his also fixes issues when
connections fail in the MiqPglogical class, and the methods called in
the rescue statement don't exist in the MiqPglogical, but do in the
PglogicalSubscription class.

This adds the `ConnectionHandling` module to share methods further
between the two classes.

Also, the `delegate` was removed since we want to also wrap that in
`.with_connection_error_handling`, and wrapping the `.pglogical` method
itself doesn't make sense.