Fix key collision with Thread[:current_request]
Now that we've moved to manageiq-loggers, the ContainerLogger has code
that expects that if Thread[:current_request] is set, it will be a
request object that responds to :request_id.  The
RequestStartedOnMiddleware, however, happens to use the same key and
stores only the request path, leading to `undefined method request_id
for "/path/to/wherever"`

Since this particular key's usage is isolated to the middleware, we can
name it whatever we want, so this commit chooses a name that is less
likely to collide.