Fix cc_test_reporter placed in wrong directory
When downloading the cc_test_reporter for the plugins the binary is put
in the spec/manageiq directory not in the plugin's root directory.

This leads to:
```
after_script
bin/ci/after_script
bin/ci/after_script: line 3: ./cc-test-reporter: No such file or directory
```

The builds don't fail because after_script failures don't fail the build
but it still fails to report the codeclimate results.