If a system_uid already exists, update it, otherwise create it
Fixes #21067

Previously, we would were reliant on the run_single_worker.rb script to
cleanly exit, and delete it's row.  If that script was terminated abnormally
the row would remain, and the next invocations of the script would create a
new row with the same system_uid value.  This leads to orphaned rows that
cause sync_workers to continually think we're out of sync.