Fix Amazon-specific provider_tag_mapping_spec
Due to a change in ManageIQ/manageiq-providers-amazon#684, flavors are
now in the list of objects in the inventory collection. This code
however had hardcoded indexes into the collections list, which broke
because of that addition. Instead, this changes the lookup to find the
correct object type making it more resilient to changes.