Lower the cpu request due to the simpler liveness probe
The liveness probe used to be in ruby and did more than it needed:
loaded some of the ruby / activesupport libraries it didn't really need, parsed
time needlessly, parsed options that were always the same, etc.

It was rewritten below to be more performant so we can now request less cpu.

https://github.com/ManageIQ/manageiq/pull/21079
https://github.com/ManageIQ/manageiq-pods/pull/688