Adjust httpd tests to account for ASCII-8BIT headers
Rails ensures that external headers are in ASCII-8BIT encoding. This
adjusts the specs to ensure that those headers are in that encoding as
they make their way through the code. This change is in anticipation of
errors that occur when Unicode characters are in headers, but in
ASCII-8BIT encoding, causing downstream errors.

Additionally, ensure that the username passed into authenticate is the
exact same value as the X-Remote-User header.