Allow ActsAsStiLeafClass for MiqWorker subclassing
If a provider inherits another provider and uses ActsAsStiLeafClass on
any of the worker records it was causing MiqWorkerType.seed to not
skip the inherited worker classes due to there being subclases.

Fix this by adding a `:concrete_subclasses` method to `MiqWorker`
similar to `ExtManagementSystem`

This wasn't needed for K8s -> OCP/EKS because those providers didn't
subclass the worker class only the runner class.  Openstack however has
some logic in the worker class which is useful and shouldn't have to be
copied into the subclassed provider workers.