[ActsAsMiqSet] Convert to miq_set_memberships
- Adds the MiqSetMembership class (join table)
- Removes RelationshipMixin from ActsAsMiqSet/ActsAsMiqSetMember
- Adds related scopes for ActsAsMiqSet/ActsAsMiqSetMember

* * *

The original approach for this effort was to create a `MiqSet` class
that would be used as the `base_class` for the rest of the models that
used to call `acts_as_miq_set_member`. However, the default behavior of
`ActiveRecord` when dealing with a polymorphic association is to set the
value of the `_type` column to `.polymorphic_name`:

    module ActiveRecord
      module Associations
        class BelongsToPolymorphicAssociation < BelongsToAssociation #:nodoc:

          private
            def replace_keys(record)
              super
              owner[reflection.foreign_type] = record ? record.class.polymorphic_name : nil
            end

Which by default for that is `base_class.name`:

    module ActiveRecord
      module Inheritance

        module ClassMethods

          def polymorphic_name
            base_class.name
          end

Classes like `MiqAlertSet` also make use of the `AssignmentMixin`, which
uses the `Tag`/`Taggings` classes to handle assignments.  When the
subclass approach was used, this broke many pieces of that as it would
save the `taggable_type` column as `"MiqSet"`, but try and find based on
the type of `"MiqAlertSet"`

With `acts_as_miq_set`, it doesn't subclass anything, just shares a
table, so it would be saved as `self.class.name`.  A lot of code relies
on this behavior, many other classes follow this properly, and it would
probably require a migration to fix old records if we were to change it
now.

By retaining the use of `ActsAsMiqSet`, this preserves with previous
behavior without mucking with the cases with `Tags` (Vms for example)
where this is done using normal `ActiveRecord` conventions.  A further
migration would also probably be required to do this correctly, so that
has been deferred for now.

Note:  With `MiqSet` not being a proper model, the individual `belongs_to`
associations needed to be defined for each `ActsAsMiqSetMember`.  This was best
done in the `ActsAsMiqSetMember` module instead of individually defining them
on the `MiqSetMembership` model