[CustomActionsMixin] Fixes for MiqSetMembership
Mostly converting the previous queries that existed to an AR based
version.

Notably:  The `uniq.flatten` needed to be switched out because previous
an array was returned by `.generic_button_group`, but now an
ActiveRecord::CollectionProxy is returned and if `uniq` is called first,
it doesn't do anything since each `CollectionProxy` is a uniq instance.
Doing a `flatten` first will convert the `CollectionProxy` instances to
Arrays, then the elements can properly have `.uniq` called on them.