Change belongsto2object_list to be more performant
The previous implementation walks the tree from root to leaf, but brings
back all the children at each level, filters in Ruby, then continues
querying down the tree, which is expensive. Instead, this commit does
the reverse by starting at the leaf and querying up the tree to the
root. This prevents bringing back a lot of unnecessary records, and also
takes advantage of the `path` method from ancestry which can query the
entire chain in a single query from the relationships table.

The specs have two tests that can show the difference in the number of
queries. One is for a shallow tree (ems/folder/folder), and another is
for a deeper tree (ems/folder/folder/folder/host). The number of queries
for the shallow tree dropped from 7 to 6 and the number of queries for
the deeper tree dropped from 13 to 7. The deeper you go in the tree, the
more savings will be had. Note that the specs have very few records at
each level in the hierarchy, but real world trees can have many objects
at each hierarchy. The new querying only brings back a single record
from each level, whereas the old query brings back all of the records at
each level, and this is where significant savings will be found.

Part of #21249