Refactor MiqReport::Formats to remove the constants
While there is nothing wrong with the original code, brakeman flags the
code as bad, which forces us to ignore it. This is the only ignore left
in the ignore file, so this commit refactors the code and the ignore is
no longer needed.

As a side-effect, this delay-loads the yml file, so we get some
object/memory savings in code paths where MiqReports are not needed.