[ActsAsMiqSet] Fix incorrect name validation
Previously, this had existed as it is now, but it didn't seem to make a
difference to the specs with the previous `RelationshipMixin` version
of this code.

However, it seems with this change, many of the specs fail with the
following:

  32) MiqWidget multiple groups for non-self service user contents created for one timezone per group with timezone_matters = false
      Failure/Error: factory_exists?(factory) ? super : class_from_symbol(factory).create!(*args)

      ActiveRecord::RecordInvalid:
        Validation failed: MiqWidgetSet: Name has already been taken

So this form seems to be the correct one.