[MiqEventDefinition] Fixes for MiqSet
There were a few things at play here causing a failure with:

    spec/models/miq_event_definition_spec.rb:68

1. The `let(:event_defs)` was actually now getting cached between calls
of `.seed_default_events`, or rather, the call to `memberof` call is now
a `association` on a `MiqSetMember` object, instead of being a method
that calls out to `Relationship`.  This meant that the first call to
`.memberof` would cache it with nothing, and the second call would
return the same.
2. The `.make_not_memberof` method was not transferred over when this
change to use `MiqSet` was originally created.  This simply adds that
method which is very similar to `.make_memberof`.  This didn't fail with
this originally because the above issue masked it since it always
returned no `.memberof` for the given `MiqEventDefinition`.
3. Made `:memberof` a proper association, and not just an alias.  This
means that you can do all of the methods associated with the relation,
instead of just the single method that is aliased (`#memberof`).