Prepend with_block, which was renamed from with_lock in Dalli 2.7.7
In commit: https://github.com/petergoldstein/dalli/commit/9f9c508afab263a2451f2209c4396daf98d33a1b
with_lock was renamed to with_block with a slightly different interface.
All versions since 2.7.7 have with_block now.

Additionally, we'll now detect and warn if the methods we depend on in our
prepended code doesn't exist before we try to prepend it.

This resolves a very similar deadlock with dalli and the rails autoloader
just like we did back in 2016 when the original patch was implemented:
https://github.com/ManageIQ/manageiq/pull/7334

We recently allowed dalli 2.7.8+, which caused these deadlocks to happen again:
https://github.com/ManageIQ/manageiq/pull/21265

Note, delete_session was removed upstream and is an alias of destroy_session, so
rely on the latter.